moved trails axios call to TrailsList component. Created ChooseYourTrailHome component.
