+++ frontend/src/App.jsx
@@ -1,9 +1,7 @@
 import { useState, useEffect } from 'react'
 import './App.css'
-import axios from 'axios'
 import { Route, Routes } from 'react-router-dom'
 
-import { BASE_URL } from './globals'
 import Header from './components/Header'
 import Main from './components/Main'
 import Home from './components/Home'
@@ -11,25 +9,11 @@ import TrailList from './components/TrailList'
 import TrailPage from './components/TrailPage'
 
 function App() {
-  const [trails, setTrails] = useState([])
-
-  useEffect(() => {
-    const getTrails = async () => {
-      const response = await axios.get(`${BASE_URL}/trails`)
-      setTrails(response.data)
-    }
-    getTrails()
-  }, [])
 
   return (
     <div>
       <Header/>
       <Main/>
-        <Routes>
-          <Route path = "/" element = {<Home /> } />
-          <Route path = "/trails" element = {<TrailList trails={trails}/>} />
-          <Route path = "/trails/:id" element = {<TrailPage trails={trails}/>} />
-        </Routes>
     </div>
         
   )
