+++ frontend/src/components/Home.jsx
@@ -1,3 +1,4 @@
+import ChooseYourTrail from "./ChooseYourTrailHome"
 import Footer from "./Footer";
 
 export default function Home () {
@@ -5,6 +6,7 @@ export default function Home () {
     return(
         <div>
             <h1> I am Home </h1>
+            <ChooseYourTrail />
             <Footer />
         </div>
     )
